Cache list of variables for the active data set
When changing the active data set via ActiveDataSet(),
generate the list of variables and save it to avoid computing
it on every call to Variables(), Factors(), TwoLevelFactors()
and Numeric(). This generates a dramatic speed gain when
activating menus, in particular when loading a new data set,
when there are many variables.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr@418 edb9625f-4e0d-4859-8d74-9fd3b1da38cb